scripts/requirements.txt
openpyxl>=3.1.0
supabase>=2.7.0
python-dotenv>=1.0.0
PyGithub>=2.3.0
geopandas>=0.14.0
shapely>=2.0.0
